backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchaudio
ffmpeg-python
numpy
soundfile
pyinstaller
pillow
